backend/face_fast_api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
cmake==3.29.2
colorama==0.4.6
dlib==19.24.2
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.14.0
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy @ file:///D:/bld/numpy_1707225561314/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=af5f40857bb7ceb2d3be562fa763863d2f954c104873e89b8ee370f75e047ecf
opencv-python==4.9.0.80
orjson==3.10.3
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.13.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.5.0
torch==2.3.0
torchvision==0.18.0
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
